pyproject.toml
[project]
name = "langchainagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.13",
    "fastapi>=0.135.3",
    "gunicorn>=25.3.0",
    "httpx>=0.27.0",
    "langchain>=0.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.2.0",
    "langserve[all]>=0.3.0",
    "protobuf<3.21",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.3",
    "streamlit>=1.55.0",
    "uvicorn>=0.44.0",
]
